refactor: split verify::execute helpers + restructure error.rs by concern
Two pre-existing items from the earlier review:

1. `verify::execute` (111 LOC) extracted into three helpers:
   - `collect_target_envs(service, target)` — single vs all branch
   - `compute_summary(reports)` — bucket counting
   - `emit_empty(output)` — JSON/human empty-case
   `execute` is now ~35 LOC of orchestration with the per-phase
   responsibilities visible at the top. No behavior change.

2. `src/error.rs` (1454 LOC, 4 impl blocks in one file) split into
   `src/error/` with per-concern submodules:

      src/error/
      ├── mod.rs        # ScoopError enum + Result alias + tests
      ├── display.rs    # message_in() + impl Display
      ├── code.rs       # code() — JSON error codes
      ├── suggestion.rs # suggestion_in() + suggestion()
      └── migrate.rs    # MigrationExitCode + migration_exit_code()

   git mv preserved history for the renamed file. All public API stays
   at `crate::error::ScoopError::*` (verified by the unchanged
   `pub use error::{Result, ScoopError};` in lib.rs).

   Source LOC per file is now under 200, down from 470 of impl bodies
   in the old monolithic file. mod.rs is still ~1100 lines but ~980 of
   those are the existing test module, intentionally kept co-located
   with the enum so the test reorg isn't bundled with the source
   reorg.

Verified
--------
- cargo build / cargo test --release: 772 passed (unchanged)
- cargo clippy --all-targets --all-features -- -D warnings: clean
- cargo fmt --check: clean
- cargo-machete: 0 unused dependencies
- Smoke: `scoop verify --help` works

References
----------
Pattern verified via context7 lookup of `thiserror` docs (impl blocks
across multiple modules for the same type — standard Rust idiom).
